12.2025 updated#3

+++ src/main/java/com/hong/demo/repository/UserManagementRepository.java
@@ -3,4 +3,4 @@ package com.hong.demo.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.hong.demo.domain.UserAccount;
 
-public interface UserManagementRepository extends JpaRepository<UserAccount, Long> {}
+public interface UserManagementRepository extends JpaRepository<UserAccount, Integer> {}
