+++ src/main/java/com/hong/demo/repository/UserRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.hong.demo.domain.UserAccount;
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<UserAccount, Long> {
+public interface UserRepository extends JpaRepository<UserAccount, Integer> {
 	// UserAccount findByUsername(String username);
 	Optional<UserAccount> findByUsername(String username);
 }
